Rename collector to aggregator.
