+++ internal/inventory/aggregator/database.go
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-package inventorycollector
+package aggregator
 
 import (
 	"context"
